[Add] styles to the load-more button and pass into the utils file the functions to use that
