+++ src/index.html
@@ -126,6 +126,9 @@
       </div>
    </div>
    <div class="add__more">
-      <button class="add__more-btn"> > </button>
+      <button class="add__more-btn">
+         <span>Load More</span>
+         <img src="./assets/icons/load-more.svg" alt="Load More">
+      </button>
    </div>
 </template>
